fix log messages.
